\name{allocationTable}
\alias{allocationTable}
\alias{allocationTable.redcapDbConnection}
\alias{allocationTable.redcapApiConnection}
\title{Allocation Tables for the Randomization Module}
\description{Generate allocation table for the REDCap randomization module. 
  Randomization may be stratified by other (categorical) variables in the 
  data set as well as by data access group.  Additionally, randomization 
  may be blocked to ensure balanced groups throughout the allocation.}

\usage{
\method{allocationTable}{redcapDbConnection}(rcon, random, strata=NULL, 
                                                group=NULL, dag_id=NULL, 
                                                replicates, block.size, 
                                                seed=NULL, 
                                                proj=NULL, ...)
\method{allocationTable}{redcapApiConnection}(rcon, random, strata=NULL, 
                                                group=NULL, dag_id=NULL, 
                                                replicates, block.size, 
                                                seed=NULL, 
                                                proj=NULL, ...)
}


\arguments{
  \item{rcon}{A REDCap connection object as generated by \code{redcapConnection}}
  \item{random}{The field name to be randomized.}
  \item{strata}{Field names by which to stratify the randomization.}
  \item{group}{A field name giving a group by which randomization should be 
    stratified.  This could also be listed in \code{strata}, but the argument
    is provided to remain consistent with the REDCap user interface.}
  \item{dag_id}{Data Access Group IDs.  See the package wiki for instructions
    on how to get the ID's. (They cannot currently be accessed via the API)}
  \item{replicates}{The number of randomizations to perform within each stratum}
  \item{block.size}{Block size for the randomization.  Blocking is recommended
    to ensure balanced groups throughout the randomization}
  \item{seed}{At least one value is required.  If only one value is given, 
    it will be converted to a vector with length equal to the number of strata.
    Values will be incremented by 100 to provide independent randomizations.
    This may also have length equal to the number of strata.}
  \item{proj}{A \code{redcapProjectInfo} object.}
  \item{...}{Arguments to be passed to other methods}
}

\details{
\code{block.size} must be a multiple of the number of groups in the randomized
variable.  If the given \code{block.size} is inappropriate, it will be increased 
to match.  For instance, if there are 4 options and \code{block.size = 6}, 
the randomization will not have equal probability for each option.  
In this case, \code{block.size} will be increased to 8 to give equal 
probability to each of the options.

\code{replicates} may also be increased slightly to ensure all blocks are complete.
Any changes to \code{replicates} or \code{block.size} are always to increase to
the next smallest appropriate number to ensure that the sampling allocation is 
larger that specified. 
}

\author{Benjamin Nutter}

\references{
More instruction on using \code{redcapAPI} to produce allocation tables is
on the package wiki:
\url{https://github.com/nutterb/redcapAPI/wiki/Randomization-Module}

Please refer to your institution's API documentation.

Additional details on API parameters are found on the package wiki at
\url{https://github.com/nutterb/redcapAPI/wiki/REDCap-API-Parameters}
}

\keyword{ methods }
