% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{constants}
\alias{constants}
\alias{REDCAP_SYSTEM_FIELDS}
\title{Constants for Use in the redcapAPI Package}
\format{
An object of class \code{character} of length 4.
}
\usage{
REDCAP_SYSTEM_FIELDS
}
\description{
Constants that are used inside the package
}
\keyword{datasets}
