% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFrameToString.R
\name{dataFrameToString}
\alias{dataFrameToString}
\title{Convert Data Frame to String Format}
\usage{
dataFrameToString(data)
}
\arguments{
\item{data}{\code{data.frame}. Each element in the data frame must
be an atomic vector.}
}
\description{
Convert Data Frame to String Format
}
\details{
Converts a data frame to the appropriate string format 
  to facility importing records, arms, events, etc.
}
