% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteFiles.R
\name{deleteFiles}
\alias{deleteFiles}
\alias{delete_files}
\title{Delete A File}
\usage{
deleteFiles(
  rcon,
  record,
  field,
  event = character(0),
  repeat_instance = 1,
  error_handling = getOption("redcap_error_handling", "error"),
  config = list(),
  api_param = list()
)

delete_files(
  rcon,
  record,
  field,
  event = character(0),
  repeat_instance = 1,
  error_handling = getOption("redcap_error_handling", "error"),
  config = list(),
  api_param = list()
)
}
\arguments{
\item{rcon}{A REDCap connection object as generated by 
\code{redcapConnection}.}

\item{record}{\code{numeric(1)} or \code{character(1)}. The record ID 
with the file to be deleted.}

\item{field}{\code{character(1)} the name of the field that contains the file.}

\item{event}{\code{character} with maximum length of one.  The unique event 
name - only for longitudinal projects}

\item{repeat_instance}{\code{numeric(1)} with maximum length of one. Only for 
projects with repeating instruments/events. The repeat instance number 
of the repeating event (if longitudinal) or the repeating instrument 
(if classic or longitudinal). Default value is '1'.}

\item{error_handling}{An option for how to handle errors returned by the API.
see \code{\link{redcapError}}}

\item{config}{\code{list} Additional configuration parameters to pass to 
\code{\link[httr]{POST}}. These are appended to any parameters in 
\code{rcon$config}.}

\item{api_param}{\code{list} Additional API parameters to pass into the
body of the API call. This provides users to execute calls with options
that may not otherwise be supported by \code{redcapAPI}.}
}
\value{
The number of deleted records.
}
\description{
This method allows you to remove a document that has been 
attached to an individual record for a File Upload field. Please note 
that this method may also be used for Signature fields (i.e. File Upload 
fields with 'signature' validation type).
}
\section{REDCap API Documentation}{

This method allows you to remove a document that has been attached to an 
individual record for a File Upload field. Please note that this method 
may also be used for Signature fields (i.e. File Upload fields with 
'signature' validation type).
}

\section{REDCap Version}{

At least 8.1.17+ (and likely some earlier versions)
}

\references{
Please refer to your institution's API documentation.
}
