% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_recapProjectInfo.R
\name{deprecated_redcapProjectInfo}
\alias{deprecated_redcapProjectInfo}
\alias{redcapProjectInfo}
\alias{redcapProjectInfo.redcapDbConnection}
\alias{redcapProjectInfo.redcapApiConnection}
\title{Deprecated Functions}
\usage{
redcapProjectInfo(rcon, date = TRUE, label = TRUE, meta_data = TRUE,
  users = TRUE, instruments = TRUE, events = TRUE, arms = TRUE,
  mappings = TRUE, version = TRUE, ...)

\method{redcapProjectInfo}{redcapDbConnection}(rcon, date = TRUE,
  label = TRUE, meta_data = TRUE, users = TRUE, instruments = TRUE,
  events = TRUE, arms = TRUE, mappings = TRUE, version = TRUE, ...)

\method{redcapProjectInfo}{redcapApiConnection}(rcon, date = TRUE,
  label = TRUE, meta_data = TRUE, users = TRUE, instruments = TRUE,
  events = TRUE, arms = TRUE, mappings = TRUE, version = TRUE, ...,
  v.number = "")
}
\arguments{
\item{rcon}{A REDCap connection object as generated by \code{redcapConnection}}

\item{date}{Logical. If \code{TRUE}, user expiration dates are converted to 
\code{POSIXct} objects.}

\item{label}{Logical.  If \code{TRUE}, the user form permissions are 
converted to labelled factors.}

\item{meta_data}{Logical.  Indicates if the meta data (data dictionary) 
should be exported.}

\item{users}{Logical. Indicates if the users table should be exported.}

\item{instruments}{Logical. Indicates if the instruments table should be exported.}

\item{events}{Logical. Indicates if the event names should be exported.}

\item{arms}{Logical. Indicates if the arms table should be exported.}

\item{mappings}{Logical. Indicates if the form-event mappings should 
be exported.}

\item{version}{Indicates if the REDCap version number should be exported.
Only applicable in REDCap 6.0.0 and higher.}

\item{...}{Arguments to be passed to other methods}

\item{v.number}{A character string given the desired version number should the
API method not be available.}
}
\description{
The \code{redcapProjectInfo} function has been deprecated to avoid
  confusion with the API method now executed by \code{exportProjectInformation}.
  The replacement function is \code{\link{exportBundle}}.
}
