% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportArms.R
\name{exportArms}
\alias{exportArms}
\alias{export_arms}
\title{Export the Arms for a Project}
\usage{
exportArms(
  rcon,
  arms = numeric(0),
  error_handling = getOption("redcap_error_handling", "error"),
  config = list(),
  api_param = list()
)

export_arms(
  rcon,
  arms = numeric(0),
  error_handling = getOption("redcap_error_handling", "error"),
  config = list(),
  api_param = list()
)
}
\arguments{
\item{rcon}{A REDCap connection object as generated by \code{redcapConnection}.}

\item{arms}{Either \code{numeric} or \code{character}. The arm numbers to retrieve.}

\item{error_handling}{An option for how to handle errors returned by the API.
see \code{\link{redcapError}}}

\item{config}{\code{list} Additional configuration parameters to pass to 
\code{\link[httr]{POST}}. These are appended to any parameters in 
\code{rcon$config}.}

\item{api_param}{\code{list} Additional API parameters to pass into the
body of the API call. This provides users to execute calls with options
that may not otherwise be supported by \code{redcapAPI}.}
}
\value{
Returns a data frame with two columns

\itemize{
  \item{\code{arm_num} }{The arm number}
  \item{\code{name} }{The arm's descriptive name}
}
}
\description{
This function allows you to export the Arms for a project
  Note: this only works for longitudinal projects
}
\details{
It is not sufficient to make the project a longitudinal project. The
project must satisfy one of two conditions: 1) have at least two arms and
one event defined in each arm; or 2) have one arm and at least two events defined. If 
neither of these conditions are satisfied, the API will return a message
such as \code{ERROR: You cannot export arms for classic projects}, an 
error message that isn't as descriptive of the nature of the problem as 
we might like.
}
\section{REDCap API Documentation}{

This function allows you to export the Arms for a project

NOTE: this only works for longitudinal projects.
}

\section{REDCap Version}{

5.8.2+
}

\section{Known REDCap Limitations}{

In versions earlier than 5.9.15, providing a value to the \code{arms} argument
had no effect and the entire data frame of arms is returned.

This was fixed in version 5.9.15.
}

\references{
Please refer to your institution's API documentation.
}
\author{
Benjamin Nutter
}
