% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exportEvents.R
\name{exportEvents}
\alias{exportEvents}
\alias{exportEvents.redcapApiConnection}
\alias{exportEvents.redcapDbConnection}
\title{Export the Events for a Project}
\usage{
exportEvents(rcon, ...)

\method{exportEvents}{redcapDbConnection}(rcon, arms = NULL, ...)
}
\arguments{
\item{rcon}{A REDCap connection object as generated by \code{redcapConnection}.}

\item{...}{Arguments to be passed to other methods.}

\item{arms}{A numeric vector or arm numbers to retrieve. In REDCap 6.5.0,
using this argument results in an empty data frame being returned.}
}
\description{
Retrieve a data frame giving the users, expiration dates,
and data access privileges for each user.
}
\details{
From the REDCap API Documentation:

This function allows you to export the Arms for a project

NOTE: this only works for longitudinal projects.

It is not sufficient to make the project a longitudinal project. The
project must satisfy one of two conditions: 1) have at least two arms and
one event defined; or 2) have one arm and at least two events defined. If
neither of these conditions are satisfied, the API will return a message
such as \code{ERROR: You cannot export arms for classic projects}, an
error message that isn't as descriptive of the nature of the problem as
we might like.
}
\section{REDCap Version}{

6.5.0+ (perhaps earlier, need to confirm its introduction)
}
\examples{
\dontrun{
#*** Note: I cannot provide working examples without
#*** compromising security.  Instead, I will try to
#*** offer up sample code with the matching results
}
}
\author{
Benjamin Nutter
}
\references{
Please refer to your institution's API documentation.

Additional details on API parameters are found on the package wiki at
\url{https://github.com/nutterb/redcapAPI/wiki/REDCap-API-Parameters}
}

