% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportInstrumentEventMappings.R
\name{exportInstrumentEventMappings}
\alias{exportInstrumentEventMappings}
\alias{export_instrument_event_mappings}
\title{Export Instrument-Event Mappings}
\usage{
exportInstrumentEventMappings(
  rcon,
  arms = numeric(0),
  error_handling = getOption("redcap_error_handling", "error"),
  config = list(),
  api_param = list()
)

export_instrument_event_mappings(
  rcon,
  arms = numeric(0),
  error_handling = getOption("redcap_error_handling", "error"),
  config = list(),
  api_param = list()
)
}
\arguments{
\item{rcon}{A REDCap connection object as generated by \code{redcapConnection}.}

\item{arms}{Either \code{numeric} or \code{character}. The arm numbers to retrieve.}

\item{error_handling}{An option for how to handle errors returned by the API.
see \code{\link{redcapError}}}

\item{config}{\code{list} Additional configuration parameters to pass to 
\code{\link[httr]{POST}}. These are appended to any parameters in 
\code{rcon$config}.}

\item{api_param}{\code{list} Additional API parameters to pass into the
body of the API call. This provides users to execute calls with options
that may not otherwise be supported by \code{redcapAPI}.}
}
\value{
Returns a data frame with two columns

\itemize{
  \item{\code{arm_num} }{The Study arm number}
  \item{\code{unique_event_name} }{The name of the event}
  \item{\code{form}{The name of the form for the arm-event.}}
}
}
\description{
This method allows you to export the instrument-event mappings 
  for a project (i.e., how the data collection instruments are designated 
  for certain events in a longitudinal project).  NOTE: This only works 
  for longitudinal projects.
}
\section{REDCap API Documentation}{

This method allows you to export the instrument-event mappings for a 
project (i.e., how the data collection instruments are designated 
for certain events in a longitudinal project).

NOTE: This only works for longitudinal projects.
}

\section{REDCap Version}{

5.8.2+
}

\section{Known REDCap Limitations}{

None.
}

\references{
Please refer to your institution's API documentation.
}
\author{
Benjamin Nutter
}
