% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exportInstruments.R
\name{exportInstruments}
\alias{exportInstruments}
\alias{exportInstruments.redcapApiConnection}
\alias{exportInstruments.redcapDbConnection}
\title{Exports the REDCap Instruments}
\usage{
exportInstruments(rcon, ...)

\method{exportInstruments}{redcapDbConnection}(rcon, ...)

\method{exportInstruments}{redcapApiConnection}(rcon, ...)
}
\arguments{
\item{rcon}{A REDCap connection object as generated by \code{redcapConnection}}

\item{...}{Arguments to be passed to other methods.}
}
\description{
Returns a data frame of instruments, names, etc.
}
\section{REDCap Version}{

6.5.0 +
}
\author{
Benjamin Nutter
}
\references{
Please refer to your institution's API documentation.

Additional details on API parameters are found on the package wiki at
\url{https://github.com/nutterb/redcapAPI/wiki/REDCap-API-Parameters}
}

