% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exportMetaData.R
\name{exportMetaData}
\alias{exportMetaData}
\alias{exportMetaData.redcapApiConnection}
\alias{exportMetaData.redcapDbConnection}
\title{Export Meta Data from a REDCap Database}
\usage{
exportMetaData(rcon, ...)

\method{exportMetaData}{redcapDbConnection}(rcon, ...)

\method{exportMetaData}{redcapApiConnection}(rcon, fields = NULL,
  forms = NULL, ...)
}
\arguments{
\item{rcon}{A REDCap connection object as generated by \code{redcapConnection.}}

\item{...}{Arguments to be passed to other methods.}

\item{fields}{A character vector of field names for which the metadata is to
be retrieved.}

\item{forms}{A character vector of forms for which the metadata is to be
retrieved. Note that if a form name is given, all of the fields on that form
will be returned, regardless of whether it is included in \code{fields} or
not.  Be careful to use the form names in the second column of the data
dictionary, and not the display names shown on the webpage.}
}
\description{
Retrieves the meta data for a REDcap database, including
field names, labels, types, formulas, etc.  This file can be used to parse
levels of factors, apply labels, and other data management tasks once the
data are retrieved
}
\details{
A record of this export is placed in the REDCap logging page,
but the file that is exported is not stored in the database.
}
\section{REDCap Version}{

5.8.2 (probably earlier, but \code{redcapAPI} wasn't written until this version)
}
\examples{
\dontrun{
#*** Note: I cannot provide working examples without
#*** compromising security.  Instead, I will try to
#*** offer up sample code with the matching results

#*** Create the connection object
rcon <- redcapConnection(url=[YOUR_REDCAP_URL], token=[API_TOKEN])

exportMetaData(rcon)
}
}
\author{
Jeffrey Horner
}
\references{
This functionality were originally developed by Jeffrey Horner in the \code{redcap} package.
\url{https://github.com/vubiostat/redcap}

Please refer to your institution's API documentation.

Additional details on API parameters are found on the package wiki at
\url{https://github.com/nutterb/redcapAPI/wiki/REDCap-API-Parameters}
}

