% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportProjectInfo.R
\name{exportProjectInfo}
\alias{exportProjectInfo}
\title{Export Project Information}
\usage{
exportProjectInfo(
  rcon,
  error_handling = getOption("redcap_error_handling", "error"),
  config = list(),
  api_param = list()
)
}
\arguments{
\item{rcon}{A REDCap connection object as generated by \code{redcapConnection}.}

\item{error_handling}{An option for how to handle errors returned by the API.
see \code{\link{redcapError}}}

\item{config}{\code{list} Additional configuration parameters to pass to 
\code{\link[httr]{POST}}. These are appended to any parameters in 
\code{rcon$config}.}

\item{api_param}{\code{list} Additional API parameters to pass into the
body of the API call. This provides users to execute calls with options
that may not otherwise be supported by \code{redcapAPI}.}
}
\value{
Returns a data frame with one row and the following columns:

\itemize{
 \item{\code{project_id}  \code{character}}
 \item{\code{project_title} \code{character}}
 \item{\code{creation_time} \code{POSIXct}}
 \item{\code{production_time} \code{POSIXct}}
 \item{\code{in_production} \code{logical}}
 \item{\code{project_language} \code{character}}
 \item{\code{purpose} \code{character}}
 \item{\code{purpose_other} \code{character}}
 \item{\code{project_notes} \code{character}}
 \item{\code{custom_record_label} \code{character}}
 \item{\code{secondary_unique_field} \code{character}}
 \item{\code{is_longitudinal} \code{logical}}
 \item{\code{surveys_enabled} \code{logical}}
 \item{\code{scheduling_enabled} \code{logical}}
 \item{\code{record_autonumbering_enabled} \code{logical}}
 \item{\code{randomization_enabled} \code{logical}}
 \item{\code{ddp_enabled} \code{logical}}
 \item{\code{project_irb_number} \code{character}}
 \item{\code{project_grant_number} \code{character}}
 \item{\code{project_pi_firstname} \code{character}}
 \item{\code{project_pi_lastname} \code{character}}
 \item{\code{display_today_now_button} \code{logical}}
 \item{\code{has_repeating_instruments_or_events} \code{logical}}
}
}
\description{
Retrieve Information about a REDCap database project.
}
\section{REDCap API Documentation}{

This method allows you to export a list of the data collection instruments 
for a project. This includes their unique instrument name as seen in the 
second column of the Data Dictionary, as well as each instrument's 
corresponding instrument label, which is seen on a project's left-hand 
menu when entering data. The instruments will be ordered according to their 
order in the project.
}

\section{REDCap Version}{

5.8.2+
}

\section{Known REDCap Limitations}{

None.
}

\references{
Please refer to your institution's API documentation.
}
\author{
Benjamin Nutter
}
