% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exportProjectInformation.R
\name{exportProjectInformation}
\alias{exportProjectInformation}
\alias{exportProjectInformation.redcapApiConnection}
\title{Exports the Project Information}
\usage{
exportProjectInformation(rcon, ...)
}
\arguments{
\item{rcon}{A REDCap connection object as generated by \code{redcapConnection}}

\item{...}{Arguments to be passed to other methods.}
}
\description{
Retrieve a data frame with the project information.
}
\details{
If this function is used in a version of REDCap that does not
  support the Export Version Number function, the character string
  \code{'Version Unknown'} is returned.
}
\section{REDCap Version}{

Sometime after 6.0.0 (exact version not yet known)
}
\author{
Benjamin Nutter
}
\references{
Please refer to your institution's API documentation.

Additional details on API parameters are found on the package wiki at
\url{https://github.com/nutterb/redcapAPI/wiki/REDCap-API-Parameters}
}

