% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportProjectXml.R
\name{exportProjectXml}
\alias{exportProjectXml}
\title{Export Project XML Data}
\usage{
exportProjectXml(
  rcon,
  filepath,
  return_metadata_only = FALSE,
  records = character(0),
  fields = character(0),
  events = character(0),
  export_survey_fields = FALSE,
  export_data_access_groups = FALSE,
  filter_logic = character(0),
  export_files = FALSE,
  meta_data = NULL,
  event_list = NULL,
  error_handling = getOption("redcap_error_handling", "error"),
  config = list(),
  api_param = list()
)
}
\arguments{
\item{rcon}{A REDCap connection object as generated by \code{redcapConnection}.}

\item{filepath}{\code{character(1)} path to the location where the XML file will be written.}

\item{return_metadata_only}{\code{logical(1)} When \code{TRUE},
returns only metadata (all fields, forms, events, and arms), 
whereas \code{FALSE} returns all metadata and also data (and optionally 
filters the data according to any of the optional parameters provided 
in the request)}

\item{records}{\code{character} or \code{numeric} vector of record names 
specifying specific records you wish to pull (by default, all 
records are pulled)}

\item{fields}{\code{character} vector of field names 
specifying specific fields you wish to pull (by default, 
all fields are pulled)}

\item{events}{\code{character} vector of unique event names that you wish 
to pull records for - only for longitudinal projects}

\item{export_survey_fields}{\code{logical(1)} specifies whether or not to 
export the survey identifier field (e.g., 'redcap_survey_identifier') or 
survey timestamp fields (e.g., instrument+'_timestamp') when surveys are 
utilized in the project. If you do not pass in this flag, it will default 
to \code{FALSE}. If set to \code{TRUE}, it will return the redcap_survey_identifier 
field and also the survey timestamp field for a particular survey when at 
least one field from that survey is being exported. NOTE: If the survey 
identifier field or survey timestamp fields are imported via API data 
import, they will simply be ignored since they are not real fields in
 the project but rather are pseudo-fields.}

\item{export_data_access_groups}{\code{logical(1)} specifies whether or not 
 to export the 'redcap_data_access_group' field when data access groups 
 are utilized in the project. If you do not pass in this flag, it will 
 default to \code{FALSE}. NOTE: This flag is only viable if the user 
 whose token is being used to make the API request is not in a data access 
 group. If the user is in a group, then this flag will revert to its
default value.}

\item{filter_logic}{\code{character} of logic text with (maximum length of 1) 
(e.g., \code{"[age] > 30"}) for filtering the data to be returned by 
this API method, in which the API will only return the records 
(or record-events, if a longitudinal project) where the logic evaluates as 
TRUE. This parameter is blank/null by default unless a value is supplied. 
Please note that if the filter logic contains any incorrect syntax, 
the API will respond with an error message.}

\item{export_files}{\code{logical(1)} \code{TRUE} will cause the XML 
returned to include all files uploaded for File Upload and Signature 
fields for all records in the project, whereas FALSE will cause all 
such fields not to be included. NOTE: Setting this option to TRUE can 
make the export very large and may prevent it from completing if the 
project contains many files or very large files.}

\item{meta_data}{\code{data.frame}, the data dictionary as returned by 
\code{exportMetaData}. If \code{NULL}, the data dictionary will be 
downloaded from the API.}

\item{event_list}{\code{data.frame}, the event list as returned by 
\code{exportEvents}.}

\item{error_handling}{An option for how to handle errors returned by the API.
see \code{\link{redcapError}}}

\item{config}{\code{list} Additional configuration parameters to pass to 
\code{\link[httr]{POST}}. These are appended to any parameters in 
\code{rcon$config}.}

\item{api_param}{\code{list} Additional API parameters to pass into the
body of the API call. This provides users to execute calls with options
that may not otherwise be supported by \code{redcapAPI}.}
}
\description{
Export Entire Project as REDCap XML File (containing metadata & data)
}
\section{REDCap API Documentation}{

This method allows you to export a list of the data collection instruments 
for a project. This includes their unique instrument name as seen in the 
second column of the Data Dictionary, as well as each instrument's 
corresponding instrument label, which is seen on a project's left-hand 
menu when entering data. The instruments will be ordered according to their 
order in the project.
}

\section{REDCap Version}{

5.8.2+
}

\section{Known REDCap Limitations}{

None.
}

\author{
Benjamin Nutter
}
