% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportRepeatingInstrumentsAndEvents.R
\name{exportRepeatingInstrumentsAndEvents}
\alias{exportRepeatingInstrumentsAndEvents}
\title{Export Repeating Instruments and Events}
\usage{
exportRepeatingInstrumentsAndEvents(
  rcon,
  error_handling = getOption("redcap_error_handling", "error"),
  config = list(),
  api_param = list()
)
}
\arguments{
\item{rcon}{A REDCap connection object as generated by \code{redcapConnection}.}

\item{error_handling}{An option for how to handle errors returned by the API.
see \code{\link{redcapError}}}

\item{config}{\code{list} Additional configuration parameters to pass to 
\code{\link[httr]{POST}}. These are appended to any parameters in 
\code{rcon$config}.}

\item{api_param}{\code{list} Additional API parameters to pass into the
body of the API call. This provides users to execute calls with options
that may not otherwise be supported by \code{redcapAPI}.}
}
\description{
Export the list of repeated instruments and events from a project.
}
\section{REDCap API Documentation}{

This method allows you to export a list of the repeated instruments and 
repeating events for a project. This includes their unique instrument name 
as seen in the second column of the Data Dictionary, as well as each 
repeating instrument's corresponding custom repeating instrument label. 
For longitudinal projects, the unique event name is also returned for 
each repeating instrument. Additionally, repeating events are returned as 
separate items, in which the instrument name will be blank/null to 
indicate that it is a repeating event (rather than a repeating instrument).
}

\section{REDCap Version}{

5.8.2+
}

\section{Known REDCap Limitations}{

None.
}

\references{
Please refer to your institution's API documentation.
}
\author{
Benjamin Nutter
}
