% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportSurveyQueueLink.R
\name{exportSurveyQueueLink}
\alias{exportSurveyQueueLink}
\alias{export_survey_queue_link}
\title{Export a Survey Queue Link for a Participant}
\usage{
exportSurveyQueueLink(
  rcon,
  record,
  error_handling = getOption("redcap_error_handling", "error"),
  config = list(),
  api_param = list()
)

export_survey_queue_link(
  rcon,
  record,
  error_handling = getOption("redcap_error_handling", "error"),
  config = list(),
  api_param = list()
)
}
\arguments{
\item{rcon}{A REDCap connection object as generated by \code{redcapConnection}.}

\item{record}{\code{character(1)} or \code{numeric(1)}, the record ID for 
which the survey link will be downloaded.}

\item{error_handling}{An option for how to handle errors returned by the API.
see \code{\link{redcapError}}}

\item{config}{\code{list} Additional configuration parameters to pass to 
\code{\link[httr]{POST}}. These are appended to any parameters in 
\code{rcon$config}.}

\item{api_param}{\code{list} Additional API parameters to pass into the
body of the API call. This provides users to execute calls with options
that may not otherwise be supported by \code{redcapAPI}.}
}
\description{
This method returns a unique Survey Queue link (i.e., a URL) 
  in plain text format for the specified record in a project that is
 utilizing the Survey Queue feature. If the user does not have 
 'Survey Distribution Tools' privileges, they will not be able to use 
 this method, and an error will be returned. If the Survey Queue feature 
 has not been enabled in the project, an error will be returned.
}
\section{REDCap API Documentation}{

This method returns a unique Survey Queue link (i.e., a URL) 
  in plain text format for the specified record in a project that is
 utilizing the Survey Queue feature. If the user does not have 
 'Survey Distribution Tools' privileges, they will not be able to use 
 this method, and an error will be returned. If the Survey Queue feature 
 has not been enabled in the project, an error will be returned.
}

\section{REDCap Version}{

5.8.2+
}

\section{Known REDCap Limitations}{

None.
}

\references{
Please refer to your institution's API documentation.
}
\author{
Benjamin Nutter
}
