% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportSurveyReturnCode.R
\name{exportSurveyReturnCode}
\alias{exportSurveyReturnCode}
\alias{export_survey_return_code}
\title{Export a Survey Return Code for a Participant}
\usage{
exportSurveyReturnCode(
  rcon,
  record,
  instrument = character(0),
  event = character(0),
  repeat_instance = 1,
  error_handling = getOption("redcap_error_handling", "error"),
  config = list(),
  api_param = list()
)

export_survey_return_code(
  rcon,
  record,
  instrument = character(0),
  event = character(0),
  repeat_instance = 1,
  error_handling = getOption("redcap_error_handling", "error"),
  config = list(),
  api_param = list()
)
}
\arguments{
\item{rcon}{A REDCap connection object as generated by \code{redcapConnection}.}

\item{record}{\code{character(1)} or \code{numeric(1)}, the record ID for 
which the survey link will be downloaded.}

\item{instrument}{\code{character} with maximium length of 1.The unique 
instrument name as seen in the second column of the Data Dictionary. 
This instrument must be enabled as a survey in the project.}

\item{event}{\code{character} with maximum legnth of 1. the unique event 
name (for longitudinal projects only).}

\item{repeat_instance}{\code{numeric(1)} (only for projects with repeating 
instruments/events) The repeat instance number of the repeating event 
(if longitudinal) or the repeating instrument (if classic or longitudinal).
Default value is '1'.}

\item{error_handling}{An option for how to handle errors returned by the API.
see \code{\link{redcapError}}}

\item{config}{\code{list} Additional configuration parameters to pass to 
\code{\link[httr]{POST}}. These are appended to any parameters in 
\code{rcon$config}.}

\item{api_param}{\code{list} Additional API parameters to pass into the
body of the API call. This provides users to execute calls with options
that may not otherwise be supported by \code{redcapAPI}.}
}
\description{
This method returns a unique Return Code in plain text format 
  for a specified record and data collection instrument 
  (and event, if longitudinal) in a project.
}
\section{REDCap API Documentation}{

If the user does not have 'Survey Distribution Tools' privileges, they will 
not be able to use this method, and an error will be returned. If the 
specified data collection instrument has not been enabled as a survey in 
the project or does not have the 'Save & Return Later' feature enabled, 
an error will be returned.
}

\section{REDCap Version}{

5.8.2+
}

\section{Known REDCap Limitations}{

None.
}

\references{
Please refer to your institution's API documentation.
}
\author{
Benjamin Nutter
}
