% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportUsers.R
\name{exportUsers}
\alias{exportUsers}
\title{Export Users}
\usage{
exportUsers(
  rcon,
  error_handling = getOption("redcap_error_handling", "error"),
  config = list(),
  api_param = list()
)
}
\arguments{
\item{rcon}{A REDCap connection object as generated by \code{redcapConnection}.}

\item{error_handling}{An option for how to handle errors returned by the API.
see \code{\link{redcapError}}}

\item{config}{\code{list} Additional configuration parameters to pass to 
\code{\link[httr]{POST}}. These are appended to any parameters in 
\code{rcon$config}.}

\item{api_param}{\code{list} Additional API parameters to pass into the
body of the API call. This provides users to execute calls with options
that may not otherwise be supported by \code{redcapAPI}.}
}
\value{
Returns a data frame

\itemize{
  \item{username} 
  \item{email}
  \item{firstname} 
  \item{lastname}
  \item{expiration} 
  \item{data_access_group} 
  \item{design}
  \item{user_rights} 
  \item{data_access_groups} 
  \item{data_export}
  \item{reports}
  \item{stats_and_charts} 
  \item{manage_survey_participants} 
  \item{calendar}
  \item{data_import_tool}
  \item{data_comparison_tool} 
  \item{logging}
  \item{file_repository} 
  \item{data_quality_create} 
  \item{data_quality_execute} 
  \item{api_export}
  \item{api_import} 
  \item{mobile_app} 
  \item{mobile_app_download_data} 
  \item{record_create} 
  \item{record_rename} 
  \item{record_delete}
  \item{lock_records_customization} 
  \item{lock_records}
  \item{lock_records_all_forms} 
  \item{forms}
}
}
\description{
This method allows you to export the list of users for a 
  project, including their user privileges and also email address, 
  first name, and last name. Note: If the user has been assigned to a 
  user role, it will return the user with the role's defined privileges.
}
\section{REDCap Version}{

5.8.2+
}

\section{Known REDCap Limitations}{

None.
}

\references{
Please refer to your institution's API documentation.
}
\author{
Benjamin Nutter
}
