% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extraction.R
\name{Extraction}
\alias{Extraction}
\alias{[.redcapFactor}
\alias{print.redcapFactor}
\title{Extraction and Assignment for \code{redcapFactor}s}
\usage{
\method{[}{redcapFactor}(x, ..., drop = FALSE)

\method{print}{redcapFactor}(x, ...)
}
\arguments{
\item{x}{an object of class \code{redcapFactor}}

\item{...}{additional arguments to pass to other methods}

\item{drop}{\code{logical}. If \code{TRUE}, unused levels are dropped.}
}
\description{
Extract elements and make assignments to \code{redcapFactor}s
}
