% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateNextRecordName.R
\name{generateNextRecordName}
\alias{generateNextRecordName}
\alias{generate_next_record_name}
\title{Generate Next Record Name}
\usage{
generateNextRecordName(
  rcon,
  error_handling = getOption("redcap_error_handling", "error"),
  config = list(),
  api_param = list()
)

generate_next_record_name(
  rcon,
  error_handling = getOption("redcap_error_handling", "error"),
  config = list(),
  api_param = list()
)
}
\arguments{
\item{rcon}{A REDCap connection object as generated by 
\code{redcapConnection}.}

\item{error_handling}{An option for how to handle errors returned by the API.
see \code{\link{redcapError}}}

\item{config}{\code{list} Additional configuration parameters to pass to 
\code{\link[httr]{POST}}. These are appended to any parameters in 
\code{rcon$config}.}

\item{api_param}{\code{list} Additional API parameters to pass into the
body of the API call. This provides users to execute calls with options
that may not otherwise be supported by \code{redcapAPI}.}
}
\value{
None.
}
\description{
To be used by projects with record auto-numbering enabled, 
  this method exports the next potential record ID for a project. 
  It generates the next record name by determining the current maximum 
  numerical record ID and then incrementing it by one.
}
\section{REDCap API Documentation}{

Note: This method does not create a new record, but merely determines 
what the next record name would be.

If using Data Access Groups (DAGs) in the project, this method accounts 
for the special formatting of the record name for users in DAGs 
(e.g., DAG-ID); in this case, it only assigns the next value for ID for 
all numbers inside a DAG. For example, if a DAG has a corresponding DAG 
number of 223 wherein records 223-1 and 223-2 already exist, then the 
next record will be 223-3 if the API user belongs to the DAG that has 
DAG number 223. (The DAG number is auto-assigned by REDCap for each DAG 
when the DAG is first created.) When generating a new record name in a DAG, 
the method considers all records in the entire project when determining 
the maximum record ID, including those that might have been originally 
created in that DAG but then later reassigned to another DAG.

Note: This method functions the same even for projects that do not have 
record auto-numbering enabled.

NOTE: This only works for longitudinal projects.
}

\section{REDCap Version}{

At least 8.1.17+ (and likely some earlier versions)
}

\references{
Please refer to your institution's API documentation.
}
