% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importArms.R
\name{importArms}
\alias{importArms}
\alias{import_arms}
\title{Import Arms To or Rename Arms In a Project}
\usage{
importArms(
  rcon,
  arms,
  arm_num = "arm_num",
  arm_name = "name",
  override = FALSE,
  error_handling = getOption("redcap_error_handling", "error"),
  config = list(),
  api_param = list()
)

import_arms(
  rcon,
  arms,
  arm_num = "arm_num",
  arm_name = "name",
  override = FALSE,
  error_handling = getOption("redcap_error_handling", "error"),
  config = list(),
  api_param = list()
)
}
\arguments{
\item{rcon}{A REDCap connection object as generated by 
\code{redcapConnection}.}

\item{arms}{\code{data.frame} with two columns.  By default, it is assumed
that the names of these columns are \code{arm_num} and \code{name}, but
these may be changed using the \code{colnames} argument. See Details for 
more about how these columns are validated.}

\item{arm_num}{\code{character(1)} The name of the column in \code{arms}
  that gives the arm number.
#' @param arm_name \code{character(1)} The name of the column in \code{arms}
  that gives the arm name.}

\item{override}{\code{logical(1)}. When \code{FALSE}, you may only add new 
or rename existing arms. When \code{TRUE}, a full overwrite is done, which 
will delete all arms and replace them with the designations in the 
\code{arms} argument.}

\item{error_handling}{An option for how to handle errors returned by the API.
see \code{\link{redcapError}}}

\item{config}{\code{list} Additional configuration parameters to pass to 
\code{\link[httr]{POST}}. These are appended to any parameters in 
\code{rcon$config}.}

\item{api_param}{\code{list} Additional API parameters to pass into the
body of the API call. This provides users to execute calls with options
that may not otherwise be supported by \code{redcapAPI}.}
}
\value{
None.
}
\description{
This method allows you to import Arms into a project or to 
  rename existing Arms in a project. To use this method, you must have API 
  Import/Update privileges AND Project Design/Setup privileges in the project.
}
\details{
The column representing the arm number must be either \code{character} 
or \code{numeric}. If it is a \code{character}, it will be coerced to 
a \code{numeric}. Before the import, the contents are tested and must be
\code{integerish}.

The column representing the arm name must be atomic. It will be converted 
to a \code{character} before the import.
}
\section{REDCap API Documentation}{

This method allows you to import Arms into a project or to rename existing 
Arms in a project. You may use the parameter \code{override = TRUE} as a 
'delete all + import' action in order to erase all existing Arms in the 
project while importing new Arms. Notice: Because of the \code{override} 
parameter's destructive nature, this method may only use 
\code{override = TRUE} for projects in Development status.

NOTE: This only works for longitudinal projects.
}

\section{REDCap Version}{

At least 8.1.17+ (and likely some earlier versions)
}

\references{
Please refer to your institution's API documentation.
}
