% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importEvents.R
\name{importEvents}
\alias{importEvents}
\alias{import_events}
\title{Import Events To or Rename Events In a Project}
\usage{
importEvents(
  rcon,
  events,
  arm_num = "arm_num",
  event_name = "event_name",
  other_cols = character(0),
  override = FALSE,
  error_handling = getOption("redcap_error_handling", "error"),
  config = list(),
  api_param = list()
)

import_events(
  rcon,
  events,
  arm_num = "arm_num",
  event_name = "event_name",
  other_cols = character(0),
  override = FALSE,
  error_handling = getOption("redcap_error_handling", "error"),
  config = list(),
  api_param = list()
)
}
\arguments{
\item{rcon}{A REDCap connection object as generated by 
\code{redcapConnection}.}

\item{events}{\code{data.frame} with at least two columns for 
\code{arm_num} and \code{event_name}.}

\item{arm_num}{\code{character(1)} The name of the column in \code{events}
with the arm number.}

\item{event_name}{\code{character(1)} The name of the column in \code{events} 
with the event name.}

\item{other_cols}{named \code{character} vector. The name of the element
will be a property recognized by the API, such as \code{unique_event_name}, 
\code{day_offset}, \code{offset_min}, or \code{offset_max}. Note: the names
are not validated--this will allow the function to work correctly if 
additional properties are added to the API in the future.  See Details
for more information about validations on properties known to be 
accepted in the API.}

\item{override}{\code{logical(1)}. When \code{FALSE}, you may only add new 
or rename existing events. When \code{TRUE}, a full overwrite is done, which 
will delete all events and replace them with the designations in the 
\code{events} argument.}

\item{error_handling}{An option for how to handle errors returned by the API.
see \code{\link{redcapError}}}

\item{config}{\code{list} Additional configuration parameters to pass to 
\code{\link[httr]{POST}}. These are appended to any parameters in 
\code{rcon$config}.}

\item{api_param}{\code{list} Additional API parameters to pass into the
body of the API call. This provides users to execute calls with options
that may not otherwise be supported by \code{redcapAPI}.}
}
\value{
None.
}
\description{
This method allows you to import Events into a project or to 
  update existing Events' attributes, such as the event name, days offset, 
  etc. The unique event name of an Event cannot be changed because it is 
  auto-generated by REDCap. To use this method, you must have API 
  Import/Update privileges AND Project Design/Setup privileges in the project.
}
\details{
The column representing the arm number must be either \code{character} 
or \code{numeric}. If it is a \code{character}, it will be coerced to 
a \code{numeric}. Before the import, the contents are tested and must be
\code{integerish}.

The column representing the event name must be atomic. It will be converted 
to a \code{character} before the import.

If \code{unique_event_name} is prescribed in \code{other_cols}, it must
be either \code{character} or \code{numeric}.

If any of \code{day_offset}, \code{offset_min}, or \code{offset_max} are
prescribed in \code{other_cols}, they must be either \code{character} 
or \code{integerish}. If they are \code{character}, it will be coerced to 
a \code{integerish}.
}
\section{REDCap API Documentation}{

This method allows you to import Events into a project or to update 
  existing Events' attributes, such as the event name, days offset, etc. 
  The unique event name of an Event cannot be changed because it is 
  auto-generated by REDCap. Please note that the only way to update an 
  existing Event is to provide the unique_event_name attribute, and if 
  the \code{unique_event_name} attribute is missing for an Event being 
  imported (when \code{override = FALSE}), it will assume it to be a 
  new Event that should be created. Notice: Because of the \code{override} 
  parameter's destructive nature, this method may only use 
  \code{override = TRUE} for projects in Development status.

NOTE: This only works for longitudinal projects.
}

\section{REDCap Version}{

At least 8.1.17+ (and likely some earlier versions)
}

\references{
Please refer to your institution's API documentation.
}
