% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importFiles.R
\name{importFiles}
\alias{importFiles}
\alias{import_files}
\title{Imports a File to REDCap to Attach to a Record}
\usage{
importFiles(
  rcon,
  file,
  record,
  field,
  event = character(0),
  overwrite = TRUE,
  repeat_instance = 1,
  bundle = getOption("redcap_bundle"),
  error_handling = getOption("redcap_error_handling", "error"),
  config = list(),
  api_param = list()
)

import_files(
  rcon,
  file,
  record,
  field,
  event = character(0),
  overwrite = TRUE,
  repeat_instance = 1,
  bundle = getOption("redcap_bundle"),
  error_handling = getOption("redcap_error_handling", "error"),
  config = list(),
  api_param = list()
)
}
\arguments{
\item{rcon}{A REDCap connection object as generated by \code{redcapConnection}}

\item{file}{\code{character(1)} giving the file path to the file to be imported.}

\item{record}{\code{integerish(1)} The record ID in which the desired file 
is to be stored.}

\item{field}{\code{character(1)} The field name in which the file is stored.}

\item{event}{\code{character(1)} The event name for the file to be stored in.
This applies only to longitudinal projects.  If the event is not supplied  
for a longitudinal project, the API will return an error}

\item{overwrite}{\code{logical(1)} When \code{FALSE}, the function checks if a 
file already exists for that record.  If a file exists, the function 
terminates to prevent overwriting.  When \code{TRUE}, no additional 
check is performed.}

\item{repeat_instance}{\code{integerish(1)} The repeat instance number of the repeating
event or the repeating instrument. When available in your instance
of REDCap, and passed as NULL, the API will assume a value of 1.}

\item{bundle}{A \code{redcapBundle} object as created by \code{exportBundle}.
This exists mostly for backward compatibility.}

\item{error_handling}{An option for how to handle errors returned by the API.
see \code{\link{redcap_error}}}

\item{config}{\code{list} Additional configuration parameters to pass to 
\code{\link[httr]{POST}}. These are appended to any parameters in 
\code{rcon$config}.}

\item{api_param}{\code{list} Additional API parameters to pass into the
body of the API call. This provides users to execute calls with options
that may not otherwise be supported by \code{redcapAPI}.}
}
\description{
A single file may be attached to a single record.  The 
  behavior of this function is consistent with the
  behavior of the API, which only allows one file to be uploaded at a time
}
\details{
The function may only import a single file
}
\author{
Benjamin Nutter
}
