% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeApiCall.R
\name{makeApiCall}
\alias{makeApiCall}
\alias{make_api_call}
\title{Make REDCap API Calls}
\usage{
makeApiCall(rcon, body = list(), config = list())

make_api_call(rcon, body = list(), config = list())
}
\arguments{
\item{rcon}{\code{redcapApiConnection} object.}

\item{body}{\code{list} List of parameters to be passed to \code{\link[httr]{POST}}'s 
\code{body} argument}

\item{config}{\code{list} A list of options to be passed to \code{\link[httr]{POST}}.
These will be appended to the \code{config} options included in the 
\code{rcon} object.}
}
\description{
Constructs and executes API calls to the REDCap API. These
  are left deliberately abstract in order to be flexible enough to 
  support the \code{redcapAPI} functions, but also allow users to 
  execute calls for new REDCap features that are not yet implemented.
}
\section{Functional Requirements}{

\enumerate{
 \item Throw an error if \code{rcon} is not a \code{redcapApiConnection}
 \item Throw an error if \code{body} is not a list.
 \item Throw an error if \code{config} is not a list.
}
}

\references{
Please refer to your institution's API documentation.
}
\author{
Benjamin Nutter
}
