% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRedcapFactor.R
\name{makeRedcapFactor}
\alias{makeRedcapFactor}
\alias{makeRedcapYN}
\alias{makeRedcapCheckbox}
\title{Factors from REDCap Variables}
\usage{
makeRedcapFactor(x, coding, factors, var_name)

makeRedcapYN(x, factors)

makeRedcapCheckbox(x, suffix, coding, factors, checkboxLabels)
}
\arguments{
\item{x}{An \code{atomic} vector.}

\item{coding}{\code{character(1)} with the coding to be applied to the 
factor.}

\item{factors}{\code{logical(1)}. Determines if the result will have the
\code{factor} class.}

\item{var_name}{\code{character(1)}. The name of the variable being converted.}

\item{suffix}{\code{character(1)}, the suffix of the checkbox variable.}

\item{checkboxLabels}{\code{logical(1)}. Determines the behavior for how 
checkbox labels are applied to the data.}
}
\description{
Convert variables to objects with the class \code{redcapFactor}.
  These factors have attributes for the levels and the labels from the 
  data dictionary.  This allows for the factor to be converted to and 
  from its original coding.  This can be useful if the levels are 
  non-sequential numeric values that have meaning, such as scores on 
  a quality of life survey.
}
\section{Functional Requirements}{


\enumerate{
 \item Return an object that has class \code{redcapFactor} 
 \item When \code{factors = TRUE} the object returns also inherits \code{factor}
 \item Throw an error if \code{x} is not an atomic vector
 \item Throw an error if \code{coding} is not a \code{character(1)}
 \item Throw an error if \code{factors} is not a \code{logical(1)}
 \item Throw an error if \code{suffix} is not a \code{character(1)}
 \item Throw an error if \code{checkboxLabels} is not a \code{logical(1)}
 \item Throw an error if \code{var_name} is not a \code{character(1)}
}
}

