\name{redcap-package}
\alias{redcap-package}
\docType{package}
\title{
Access data, meta data, and files from REDCap using the API
}
\description{
REDCap API functions are made available to the user.  Data may be formatted consistent with the data dictionary for 
analysis-ready data.
}
\details{
\tabular{ll}{
Package: \tab redcap\cr
Type: \tab Package\cr
Version: \tab 0.9\cr
Date: \tab 2014-08-01\cr
License: \tab gpl\cr
LazyLoad: \tab yes\cr
}

}
\author{
Jeffery Horner, Jeremy Stevens, Will Gray, Benjamin Nutter

Maintainer: Benjamin Nutter <nutterb@ccf.org>

}
\keyword{ package }
