% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcapConnection.R
\name{redcapConnection}
\alias{redcapConnection}
\alias{redcap_connection}
\alias{print.redcapApiConnection}
\title{Connect to a REDCap Database}
\usage{
redcapConnection(url = getOption("redcap_api_url"), token, config = list())

redcap_connection(url = getOption("redcap_api_url"), token, config = list())

\method{print}{redcapApiConnection}(x, ...)
}
\arguments{
\item{url}{\code{character(1)} The URL for a REDCap API connection.
Check your institution's REDCap documentation for this address.}

\item{token}{\code{character(1)} REDCap API token. It is expected that 
this will be exactly 32 characters (after removing leading and 
trailing whitespace)}

\item{config}{\code{list} of configuration parameters to pass to 
\code{\link[httr]{config}}. This allows the 
user to set additional configurations for the API calls, such as 
certificates, ssl version, etc.}

\item{x}{A \code{redcapApiConnection} object to be printed.}

\item{...}{Additional arguments to pass to other methods.}
}
\description{
Creates an object of class \code{redcapApiConnection} for 
using the REDCap API
}
\details{
To obtain an API token for a project, do the following:\cr
Enter the 'User Right' section of a project\cr
Select a user\cr
Check the box for 'API Data Export' or 'API Data Import,' as appropriate. 

Tokens are specific to a project, and a token must be created for each 
project for which you wish to use the API.

Additional Curl option can be set in the \code{config} argument.  See the documentation
for \code{\link[httr]{config}} and \code{\link[httr]{httr_options}} for more Curl options.
}
\section{Functional Requirements}{

\enumerate{
 \item Return an object of class \code{redcapApiConnection}
 \item Leading and trailing white space are removed from \code{token}
 \item Throw an error if \code{url} is not a \code{character(1)}
 \item Throw an error if \code{token} is not a \code{character(1)}
 \item Throw an error if \code{token} does not have either 32 or 64 characters
 \item Throw an error if \code{config} is not a \code{list}
}
}

\references{
This functionality were originally developed by Jeffrey Horner in the 
\code{redcap} package.
\url{https://github.com/vubiostat/redcap}
}
\author{
Jeffrey Horner, Benjamin Nutter
}
