\name{redcapProjectInfo}
\alias{redcapProjectInfo}
\alias{redcapProjectInfo.redcapDbConnection}
\alias{redcapProjectInfo.redcapApiConnection}
\title{Export a Project's Supplemental Information}
\description{The meta data, users, arms, events, and mappings can be used to do some validation during data
  exports and imports.  In order to reduce calls to the API, these table can be stored in the options and 
  referenced as needed.}

\usage{
\method{redcapProjectInfo}{redcapDbConnection}(rcon, date=TRUE, label=TRUE, ...)
\method{redcapProjectInfo}{redcapApiConnection}(rcon, date=TRUE, label=TRUE, ...)
}


\arguments{
  \item{rcon}{A REDCap connection object as generated by \code{redcapConnection}}
  \item{date}{Logical. If \code{TRUE}, user expiration dates are converted to \code{POSIXct} objects.}
  \item{label}{Logical.  If \code{TRUE}, the user form permissions are converted to labelled factors.}
  \item{...}{Arguments to be passed to other methods}
}

\details{
The project information is stored in the option \code{redcap_project_info}.  If the project is not longitudinal, the 
events, arms, and event-form mappings elements will be assigned character vectors instead of data frames.
}

\author{Benjamin Nutter}

\examples{
\dontrun{
> #*** Note: I cannot provide working examples without
> #*** compromising security.  Instead, I will try to 
> #*** offer up sample code with the matching results
> 
> 
> #*** Create the connection object
> rcon <- redcapConnection(url=[YOUR_REDCAP_URL], token=[API_TOKEN])
> 
> redcapProjectInfo(rcon)
}
}

\keyword{ methods }
