% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TestCase01_Data}
\alias{TestCase01_Data}
\title{REDCap Test Case 01: Formatted Data}
\format{
A data frame with 100 rows and 36 columns.
}
\usage{
TestCase01_Data
}
\description{
These data are the formatted data for a data set consisting 
of a cross-sectional, single center, single form, single arm 
design.  The data are what should be an expected result from 
a call to \code{exportRecords} with the default values.
}
\keyword{datasets}
