% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TestCase01_MetaData}
\alias{TestCase01_MetaData}
\title{REDCap Test Case 01: Meta Data / Data Dictionary}
\format{
A data frame with 31 rows and 18 columns.
}
\source{
Randomly generated and meaningless data.
}
\usage{
TestCase01_MetaData
}
\description{
These data are the data dictionary for a data set consisting 
of a cross-sectional, single center, single form, single arm 
design. This is available to assist with testing.  The data 
dictionary is designed to include at least one column of 
each data type.
}
\keyword{datasets}
