% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkboxSuffixes.R
\name{checkboxSuffixes}
\alias{checkboxSuffixes}
\title{Checkbox Suffixes}
\usage{
checkboxSuffixes(fields, meta_data, version)
}
\arguments{
\item{fields}{\code{character} The current field names of interest}

\item{meta_data}{\code{data.frame}. The meta data data frame.}

\item{version}{\code{character(1)} The REDCap version number.}
}
\description{
Checkbox variables return one vector of data for each option defined
  in the variable.  The variables are returned with the suffix \code{___[option]}.
  \code{exportRecords} needs these suffixes in order to retrieve all of the 
  variables and to apply the correct labels.
}
\section{Functional Requirements}{

\enumerate{
 \item Return a list with two vectors of names and labels.
 \item Throw an error if \code{fields} is not a \code{character}
 \item Throw an error if \code{meta_data} is not a \code{data.frame}
 \item Throw an error if \code{version} is not a \code{character(1)}
}
}

