\name{deleteFiles}
\alias{deleteFiles}
\alias{deleteFiles.redcapDbConnection}
\alias{deleteFiles.redcapApiConnection}
\title{Deletes a File attached to a Record}
\description{This function allows you to remove a document that has been attached to an individual record}

\usage{
\method{deleteFiles}{redcapDbConnection}(rcon, record, field, event, ...)
\method{deleteFiles}{redcapApiConnection}(rcon, record, field, event, ..., 
    meta_data=getOption("redcap_project_info")$meta_data,
    events_list = getOption("redcap_project_info")$events)
}


\arguments{
  \item{rcon}{A REDCap connection object as generated by \code{redcapConnection}}
  \item{record}{The record ID in which the desired file is stored. Must be length 1.}
  \item{field}{The field name in which the file is stored. Must be length 1.}
  \item{event}{The event name for the file.  Must be length 1.  This applies only to longitudinal projects.  
    If the event is not
    supplied for a longitudinal project, the API will return the message "Error: Bad Request"}
  \item{meta_data}{A data frame of the project meta_data.  This is used to do parameter checking prior to making
    the call to the API.  It is strongly recommended that you run \code{redcapProjectInfo} prior to exporting
    records.  If necessary, an API call will be made to retrieve this table.}
  \item{events_list}{A data frame with the list of events in a longitudinal project.  This is used to do parameter
    checking prior to making the call to the API.  It is strongly recommended that you run \code{redcapProjectInfo}
    prior to exporting records.  If necessary, an API call will be made to retrieve this table.}
   \item{...}{Arguments to be passed to other methods}
}

\author{Benjamin Nutter}

\examples{
\dontrun{
> #*** Note: I cannot provide working examples without
> #*** compromising security.  Instead, I will try to 
> #*** offer up sample code with the matching results
> 
> #*** Create the connection object
> rcon <- redcapConnection(url=[YOUR_REDCAP_URL], token=[API_TOKEN])
> 
>
> #* Delete a file
> deleteFiles(rcon, record=1, field="file_upload", event="event_1_arm_1")
The file was successfully deleted
>
}
}

\keyword{ methods }
