% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportInstruments.R
\name{exportInstruments}
\alias{exportInstruments}
\title{Export Instruments}
\usage{
exportInstruments(
  rcon,
  error_handling = getOption("redcap_error_handling", "error"),
  config = list(),
  api_param = list()
)
}
\arguments{
\item{rcon}{A REDCap connection object as generated by \code{redcapConnection}.}

\item{error_handling}{An option for how to handle errors returned by the API.
see \code{\link{redcapError}}}

\item{config}{\code{list} Additional configuration parameters to pass to 
\code{\link[httr]{POST}}. These are appended to any parameters in 
\code{rcon$config}.}

\item{api_param}{\code{list} Additional API parameters to pass into the
body of the API call. This provides users to execute calls with options
that may not otherwise be supported by \code{redcapAPI}.}
}
\value{
Returns a data frame with two columns

\itemize{
  \item{\code{instrument_name} }{The internal name of the instrument}
  \item{\code{instrument_label} }{The display name of the instrument}
}
}
\description{
Retrieve the names of instruments stored in the REDCap Project.
}
\section{REDCap API Documentation}{

This method allows you to export a list of the data collection instruments 
for a project. This includes their unique instrument name as seen in the 
second column of the Data Dictionary, as well as each instrument's 
corresponding instrument label, which is seen on a project's left-hand 
menu when entering data. The instruments will be ordered according to their 
order in the project.
}

\section{REDCap Version}{

5.8.2+
}

\section{Known REDCap Limitations}{

None.
}

\references{
Please refer to your institution's API documentation.
}
\author{
Benjamin Nutter
}
