% Generated by roxygen2 (4.0.2): do not edit by hand
\name{exportInstruments}
\alias{exportInstruments}
\alias{exportInstruments.redcapApiConnection}
\alias{exportInstruments.redcapDbConnection}
\title{Export Instrument names}
\usage{
exportInstruments(rcon, ...)

\method{exportInstruments}{redcapDbConnection}(rcon, ...)

\method{exportInstruments}{redcapApiConnection}(rcon, ...)
}
\arguments{
\item{rcon}{A REDCap connection object as generated by \code{redcapConnection}}

\item{...}{Arguments to be passed to other methods}
}
\description{
Retrieve a data frame giving the instrument names and labels in a project.
}
\details{
This function allows you to export a list of the data collection
instruments for a project. This includes their unique instrument name as
seen in the second column of the Data Dictionary, as well as each
instrument's corresponding instrument label, which is seen on a project's
left-hand menu when entering data. The instruments will be ordered
according to their order in the project.

This function was introduced to the API in version 5.9.  An error will be
returned if called to an instance earlier than 5.9.
}
\author{
Benjamin Nutter
}
\references{
Please refer to your institution's API documentation.

Additional details on API parameters are found on the package wiki at
\url{https://github.com/nutterb/redcapAPI/wiki/REDCap-API-Parameters}
}

