% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportMetaData.R
\name{exportMetaData}
\alias{exportMetaData}
\alias{export_meta_data}
\alias{exportDataDictionary}
\alias{export_data_dictionary}
\title{Export Meta Data from a REDCap Database}
\usage{
exportMetaData(
  rcon,
  fields = character(0),
  forms = character(0),
  error_handling = getOption("redcap_error_handling", "null"),
  config = list(),
  api_param = list()
)

export_meta_data(
  rcon,
  fields = character(0),
  forms = character(0),
  error_handling = getOption("redcap_error_handling", "null"),
  config = list(),
  api_param = list()
)

exportDataDictionary(
  rcon,
  fields = character(0),
  forms = character(0),
  error_handling = getOption("redcap_error_handling", "null"),
  config = list(),
  api_param = list()
)

export_data_dictionary(
  rcon,
  fields = character(0),
  forms = character(0),
  error_handling = getOption("redcap_error_handling", "null"),
  config = list(),
  api_param = list()
)
}
\arguments{
\item{rcon}{A REDCap connection object as generated by \code{redcapConnection.}}

\item{fields}{A character vector of field names for which the metadata is to 
be retrieved.}

\item{forms}{A character vector of forms for which the metadata is to be
retrieved. Note that if a form name is given, all of the fields on that form
will be returned, regardless of whether it is included in \code{fields} or 
not.  Be careful to use the form names in the second column of the data 
dictionary, and not the display names shown on the webpage.}

\item{error_handling}{An option for how to handle errors returned by the API.
see \code{\link{redcapError}}. One of \code{c("null", "error")}.}

\item{config}{Additional controls to be passed to other \code{\link[httr]{config}}.
These will be appended to the controls given in \code{rcon}.}

\item{api_param}{\code{list} Additional API parameters to pass into the
body of the API call. This provides users to execute calls with options
that may not otherwise be supported by \code{redcapAPI}.}
}
\description{
Retrieves the meta data for a REDcap database, including 
field names, labels, types, formulas, etc.  This file can be used to parse 
levels of factors, apply labels, and other data management tasks once the 
data are retrieved
}
\details{
A record of this export is placed in the REDCap logging page, 
but the file that is exported is not stored in the database.
}
\section{REDCap API Documentation}{

This function allows you to export the metadata for a project
}

\section{REDCap Version}{

5.8.2+ (and earlier, but we don't know how much earlier)
}

\section{Functional Requirements}{

\enumerate{
 \item Return a data frame with the meta data.
 \item Throw an error if \code{fields} is not a \code{character}
 \item Throw an error if \code{forms} is not a \code{character}
 \item Throw an error if \code{error_handling} is not one of \code{c("null", "error")}
}
}

\references{
This functionality was originally developed by Jeffrey Horner in the \code{redcap} package.
\url{https://github.com/vubiostat/redcap}

Please refer to your institution's API documentation.
}
\author{
Jeffrey Horner
}
