% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportReports.R
\name{exportReports}
\alias{exportReports}
\title{Export Reports}
\usage{
exportReports(
  rcon,
  report_id,
  error_handling = getOption("redcap_error_handling", "error"),
  config = list(),
  api_param = list()
)
}
\arguments{
\item{rcon}{A REDCap connection object as generated by \code{redcapConnection}.}

\item{report_id}{\code{character(1)} or \code{numeric(1)} giving the report 
ID to download.}

\item{error_handling}{An option for how to handle errors returned by the API.
see \code{\link{redcapError}}}

\item{config}{\code{list} Additional configuration parameters to pass to 
\code{\link[httr]{POST}}. These are appended to any parameters in 
\code{rcon$config}.}

\item{api_param}{\code{list} Additional API parameters to pass into the
body of the API call. This provides users to execute calls with options
that may not otherwise be supported by \code{redcapAPI}.}
}
\description{
This method allows you to export the data set of a report created on a project's 'Data Exports, Reports, and Stats' page.
}
\section{REDCap API Documentation}{

Please be aware that Data Export user rights will be applied to this API 
request. For example, if you have 'No Access' data export rights in the 
project, then the API report export will fail and return an error. And 
if you have 'De-Identified' or 'Remove all tagged Identifier fields' data 
export rights, then some data fields might be removed and filtered out 
of the data set returned from the API. To make sure that no data is 
unnecessarily filtered out of your API request, you should have 
'Full Data Set' export rights in the project.

Also, please note the the 'Export Reports' method does not make use of 
the 'type' (flat/eav) parameter, which can be used in the 'Export Records'
method. All data for the 'Export Reports' method is thus exported in flat
format. If the 'type' parameter is supplied in the API request, it will 
be ignored.
}

\section{REDCap Version}{

5.8.2+
}

\section{Known REDCap Limitations}{

None.
}

\references{
Please refer to your institution's API documentation.
}
\author{
Benjamin Nutter
}
