% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportSurveyParticipants.R
\name{exportSurveyParticipants}
\alias{exportSurveyParticipants}
\alias{export_survey_participants}
\title{Export Survey Participants}
\usage{
exportSurveyParticipants(
  rcon,
  instrument,
  event = character(0),
  error_handling = getOption("redcap_error_handling", "error"),
  config = list(),
  api_param = list()
)

export_survey_participants(
  rcon,
  instrument,
  event = character(0),
  error_handling = getOption("redcap_error_handling", "error"),
  config = list(),
  api_param = list()
)
}
\arguments{
\item{rcon}{A REDCap connection object as generated by \code{redcapConnection}.}

\item{instrument}{\code{character(1)} the unique instrument name as seen 
in the second column of the Data Dictionary. This instrument must be 
enabled as a survey in the project.}

\item{event}{\code{character}  with maximum length of 1. The unique event 
name (for longitudinal projects only).}

\item{error_handling}{An option for how to handle errors returned by the API.
see \code{\link{redcapError}}}

\item{config}{\code{list} Additional configuration parameters to pass to 
\code{\link[httr]{POST}}. These are appended to any parameters in 
\code{rcon$config}.}

\item{api_param}{\code{list} Additional API parameters to pass into the
body of the API call. This provides users to execute calls with options
that may not otherwise be supported by \code{redcapAPI}.}
}
\value{
Returns a data frame

\itemize{
  \item{\code{email} Participant e-mail address} 
  \item{\code{email_occurrence} represents the current count that the email 
    address has appeared in the list (because emails can be used more 
    than once), thus email + email_occurrence represent a unique value pair.}
  \item{\code{identifier} } 
  \item{\code{record} }
  \item{\code{invitation_sent_status} Indicates if an invitation has been sent.} 
  \item{\code{invitation_send_time} is the date/time in which the next 
    invitation will be sent, and is blank if there is no invitation that 
    is scheduled to be sent} 
  \item{\code{response_status} represents whether the participant has responded to the survey}
  \item{\code{survey_access_code}} 
  \item{\code{survey_link}} 
  \item{\code{survey_queue_link}}
}
}
\description{
Export a Survey Participant List
}
\section{REDCap Version}{

5.8.2+
}

\section{Known REDCap Limitations}{

None.
}

\references{
Please refer to your institution's API documentation.
}
\author{
Benjamin Nutter
}
