\name{exportUsers}
\alias{exportUsers}
\title{Exports the Users for a Project}
\description{Retrieve a data frame giving the users, expiration dates, and data access privileges for each user.}

\usage{exportUsers(rcon, date=TRUE, label=TRUE)}


\arguments{
  \item{rcon}{A REDCap connection object as generated by }
  \item{date}{Logical.  Indicates if the expiration date is converted to a \code{POSIXct} object.}
  \item{label}{Logical. Indicates if the data export and form access rights are converted to factors objects.}
}

\details{
From the REDCap API Documentation:

User information for the project in the format specified

Data Export: 0=no access, 2=De-Identified, 1=Full Data Set\cr
Form Rights: 0=no access, 2=read only, 1=view records/responses and edit records (survey responses are read-only), 
3 = edit survey responses

(NOTE: At this time, only a limited amount of rights-related info will be exported (expiration, data access group ID, 
data export rights, and form-level rights. However, more info about a user's rights will eventually be added to the 
Export Users API functionality in future versions of REDCap.)}

\author{Benjamin Nutter}

\keyword{ methods }
