% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportVersion.R
\name{exportVersion}
\alias{exportVersion}
\alias{export_version}
\title{Exports the REDCap Version Number}
\usage{
exportVersion(
  rcon,
  error_handling = getOption("redcap_error_handling", "null"),
  config = list(),
  api_param = list()
)

export_version(
  rcon,
  error_handling = getOption("redcap_error_handling", "null"),
  config = list(),
  api_param = list()
)
}
\arguments{
\item{rcon}{A \code{redcapConnection} object.}

\item{error_handling}{\code{character}, one of \code{c("null", "error")}. 
An option for how to handle errors returned by the API.
see \code{\link{redcapError}}}

\item{config}{\code{list}. Additional parameters to pass to 
\code{\link[httr]{POST}}.}

\item{api_param}{\code{list}. Additional parameters to pass to the 
REDCap API. This provides users to execute calls with options
that may not otherwise be supported by \code{redcapAPI}.}
}
\description{
Version numbers are returned as a character string.
  This feature is available for REDCap 6.0.0 and higher.
}
\details{
If this function is used in a version of REDCap that does not
  support the Export Version Number function, the character string
  \code{'5.12.2'} is returned. This is done solely for the convenience 
  of always returning a value that can be compared against other versions.
}
\section{REDCap API Documentation (6.5.0)}{

This method returns the current REDCap version number as plain text 
(e.g., 4.13.18, 5.12.2, 6.0.0).
}

\section{REDCap Version}{

6.0.0+
}

\section{Known REDCap Limitations}{

None
}

\references{
Please refer to your institution's API documentation.
}
\author{
Benjamin Nutter
}
