% Generated by roxygen2 (4.0.2): do not edit by hand
\name{exportVersion}
\alias{exportVersion}
\alias{exportVersion.redcapApiConnection}
\alias{exportVersion.redcapDbConnection}
\title{Exports the REDCap Version Number}
\usage{
exportVersion(rcon, ...)

\method{exportVersion}{redcapDbConnection}(rcon, ...)

\method{exportVersion}{redcapApiConnection}(rcon, ...)
}
\arguments{
\item{rcon}{A REDCap connection object as generated by \code{redcapConnection}}

\item{...}{Arguments to be passed to other methods.}
}
\description{
Version numbers are returned as a character string.
  This feature is available for REDCap 6.0.0 and higher.
}
\details{
If this function is used in a version of REDCap that does not
  support the Export Version Number function, the character string
  \code{'Version Unknown'} is returned.
}
\author{
Benjamin Nutter
}
\references{
Please refer to your institution's API documentation.

Additional details on API parameters are found on the package wiki at
\url{https://github.com/nutterbAPI/redcap/wiki/REDCap-API-Parameters}
}

