\name{exportVersion}
\alias{exportVersion}
\alias{exportVersion.redcapDbConnection}
\alias{exportVersion.redcapApiConnection}
\title{Exports the REDCap Version Number}
\description{Version numbers are returned as a character string.  This feature is available for REDCap 6.0.0 and higher.}

\usage{
\method{exportVersion}{redcapDbConnection}(rcon, ...)
\method{exportVersion}{redcapApiConnection}(rcon, ...)
}


\arguments{
  \item{rcon}{A REDCap connection object as generated by \code{redcapConnection}}
  \item{...}{Arguments to be passed to other methods}
}

\details{
If this function is used in a version of REDCap that does not support the Export Version Number function, the character
string \code{'Version Unknown'} is returned.  }

\author{Benjamin Nutter}

\references{
Please refer to your institution's API documentation.

Additional details on API parameters are found on the package wiki at
\url{https://github.com/nutterbAPI/redcap/wiki/REDCap-API-Parameters}
}
