% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fieldToVar.R
\name{fieldToVar}
\alias{fieldToVar}
\title{Convert a REDCap Data Field to an R Vector}
\usage{
fieldToVar(records, meta_data, factors = TRUE, dates = TRUE,
  checkboxLabels = FALSE)
}
\description{
Converts a field exported from REDCap into a valid R vector
}
\details{
This function is called internally by \code{exportRecords} and
  \code{exportReports}.  it is not available to the user.
}
\author{
Jeffrey Horner
}

