\name{importFiles}
\alias{importFiles}
\alias{importFiles.redcapDbConnection}
\alias{importFiles.redcapApiConnection}
\title{Imports a File to Attach to a Record}
\description{A single file may be attached to a single record.  The behavior of this function is consistent with the
    behavior of the API, which only allows one file to be uploaded at a time.}

\usage{
\method{importFiles}{redcapDbConnection}(rcon, file, record, field, event, 
    meta_data=getOption("redcap_project_info")$meta_data,
    events_list = getOption("redcap_project_info")$events, ...)
\method{importFiles}{redcapApiConnection}(rcon, record, field, event, 
    meta_data=getOption("redcap_project_info")$meta_data,
    events_list = getOption("redcap_project_info")$events, ...)
}


\arguments{
  \item{rcon}{A REDCap connection object as generated by \code{redcapConnection}}
  \item{file}{Character string giving the file path to the file to be imported.}
  \item{record}{The record ID in which the desired file is stored. Must be length 1.}
  \item{field}{The field name in which the file is stored. Must be length 1.}
  \item{event}{The event name for the file.  Must be length 1.  This applies only to longitudinal projects.  
    If the event is not
    supplied for a longitudinal project, the API will return the message "Error: Bad Request"}
  \item{meta_data}{A data frame of the project meta_data.  This is used to do parameter checking prior to making
    the call to the API.  It is strongly recommended that you run \code{redcapProjectInfo} prior to exporting
    records.  If necessary, an API call will be made to retrieve this table.}
  \item{events_list}{A data frame with the list of events in a longitudinal project.  This is used to do parameter
    checking prior to making the call to the API.  It is strongly recommended that you run \code{redcapProjectInfo}
    prior to exporting records.  If necessary, an API call will be made to retrieve this table.}    
  \item{...}{Arguments to be passed to other methods}
}

\details{
The function may only import a single file.}

\author{Benjamin Nutter}

\examples{
\dontrun{
> #*** Note: I cannot provide working examples without
> #*** compromising security.  Instead, I will try to 
> #*** offer up sample code with the matching results
> 
> #*** Create the connection object
> rcon <- redcapConnection(url=[YOUR_REDCAP_URL], token=[API_TOKEN])
> 
>
> #* Import a single file
> importFiles(rcon, "Image.jpg", record=1, field="file_upload", event="event_1_arm_1")
The file was successfully uploaded
>
}
}

\keyword{ methods }
