% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorToApiBodyList.R
\name{vectorToApiBodyList}
\alias{vectorToApiBodyList}
\title{Convert R Vector To List for the API Call Body}
\usage{
vectorToApiBodyList(vector, parameter_name)
}
\arguments{
\item{vector}{An \code{atomic} vector.}

\item{parameter_name}{\code{character(1)}. The REDCap API parameter name.}
}
\description{
Converts an R vector to a list that will be suitable for 
  \code{makeApiCall}.
}
\section{Functional Requirements}{
 
\enumerate{
 \item Return a named list. 
 \item The name of each element in the list starts with the value of \code{parameter_name}
 \item Throw an error if \code{x} is not an atomic vector
 \item Throw an error if \code{parameter_name} is not a \code{character(1)}
}
}

\author{
Benjamin Nutter with thanks to Philip Chase for showing me how 
  these need to be formatted.
}
